requirements.txt
streamlit
pandas
nltk
textblob
plotly
selenium
webdriver-manager
pillow
tweet-preprocessor
